requirements.txt
altair==5.0.1
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dacite==1.8.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
greenplum-python @ git+https://github.com/greenplum-db/GreenplumPython
htmlmin==0.1.12
huggingface-hub==0.15.1
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multimethod==1.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.23.5
packaging==23.1
pandas==2.0.2
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pgvector==0.1.8
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prettytable==3.7.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.23.1
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typeguard==2.13.3
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.6
validators==0.20.0
visions==0.7.5
wcwidth==0.2.6
wordcloud==1.9.2
ydata-profiling==4.3.0
zipp==3.15.0
